Add bangla zuktabarna related mappings
There current mappings were missing quite some bangla combined letters
(zuktabarna). This was causing some strange transilation, this commit
adds those mappings for those missing zuktabarna.

This is not a official list, but based on some primary research it is
seems to be quite reliable, so we can start with this one and update
it as soon as we find something reliable.

This also adds some bangla text for test with expected output, it
is not 100% accurate, but pretty close to the way a native will read
those out.

Source: https://www.somewhereinblog.net/blog/trivuzblog/28849694