Add an exceptional mapping to fix the tests
The mongolian mappings contains some tests that expect us to do
a proper mapping for a exception case (a city name in this case)
without having some proper mappings for those.

For now, this commit adds that whole word as exceptional to keep
our test suite green, but we will need a native to come up with
some sort of rules for those.