Add extending capability to the systems
Recently, we came across a use case where we wanted to re-use
character mappings from an another system, and this lead to the
idea that we might wanna add some extending capabilities to a
system mappings.

This commit adds that functionality, and it also keep this
mapping handling separate so if necessary then this would be
the only place we will need to update in the future.