Add support for character separator
This commit adds support for character separator, so whenever we are
specifying mappings for characters then we can also specify a newly
added key `character_separator`. This separator will be added during
the transliteration phase for each character.

By default, it is empty so not other system should be affected but
it does adds a little bit of overhead during the transliteration
phase.

Fixes #44