Implement dictionary lookup
This commit implements dictionary lookup function.

(1) `transliterate`  will now scan the input string, check against the `dictionary` hash, and replace the longest match with the lookup results.
(2) Dictionary will be included when the `inherit` option is handled.
(3) The building of characters and dictionary hash has been moved out of `transliterate`
(4) Caching of `maps` so that a map only needs to be loaded once.