kp-kor-Hang-Latn-2002: Fix a problem with punctuation (metanorma#341)
We added a test case that a string:
"구현, 글꼴, 문자 배열, 다국어 컴퓨팅."
Should become:
"Kuhyŏn, Külkkol, Munja Paeyŏl, Tagugŏ Khŏmphyuthing."
Before that commit it was:
"Kuhyŏᆫ, Külkkoᆯ, Munja Paeyŏᆯ, Tagugŏ Khŏmphyuthiᆼ."

The last rules assumed the only punctuation that is about to happen
was "-", and "," and "." weren't handled properly. Therefore this
commit replaces the whole regexp with a wider match.

There are other rules which may become problematic in the future,
that is lines from 621 to 658.