Fix Interscript for new Psych releases
Psych (the Ruby YAML library) has changed semantics of `YAML.load`
frim being `unsafe_load` by default, to being `safe_load`.

`safe_load` differs from `unsafe_load` by whitelisting a number of
classes.

We rely on the old behavior, as some maps contain date in the Date
format and Date is not in a whitelisted set.

Older versions of Psych may not support a `unsafe_load` method and
those versions are chosen for certain older Ruby versions that we
support.

The question of safety in this case is secondary, since the maps
are already executed as Ruby code.